requirements.txt
matplotlib>=3.0.0
Flask==2.3.3
Flask-CORS==4.0.0
ultralytics==8.3.0
opencv-python==4.8.0.76
numpy==1.24.3
Pillow==10.0.0
Werkzeug==2.3.7
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
torch